requirements.txt
asgiref==3.6.0
biopython==1.81
blinker==1.6.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
Django==4.2.1
django-filter==23.3
djangorestframework==3.14.0
filelock==3.12.4
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
platformdirs==3.10.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.5
Werkzeug==2.3.7
